requirements.txt
analytics-python==1.4.0
anyio==3.3.4
asgiref==3.4.1
astor==0.8.1
Babel==2.9.1
backoff==1.10.0
backports.entry-points-selectable==1.1.0
bce-python-sdk==0.8.62
bcrypt==3.2.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
colorlog==6.5.0
contextlib2==21.6.0
contextvars==2.4
cryptography==35.0.0
cycler==0.11.0
dask==2021.3.0
dataclasses==0.8
decorator==4.4.2
dill==0.3.4
distlib==0.3.3
easydict==1.9
fastapi==0.70.0
ffmpy==0.3.0
filelock==3.3.2
flake8==3.9.2
Flask==2.0.2
Flask-Babel==2.0.0
Flask-CacheBuster==1.0.0
Flask-Cors==3.0.10
Flask-Login==0.5.0
future==0.18.2
gast==0.3.3
gitdb==4.0.9
GitPython==3.1.18
gradio==2.4.1
h11==0.12.0
h5py==3.1.0
httptools==0.2.0
identify==2.3.1
idna==3.3
immutables==0.16
importlib-metadata==4.8.1
importlib-resources==5.3.0
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.2
joblib==1.1.0
kiwisolver==1.3.1
markdown2==2.4.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.6.1
monotonic==1.6
multiprocess==0.70.12.2
networkx==2.5.1
nodeenv==1.6.0
numpy==1.15.2
onnx==1.8.0
opencv-python==4.5.4.58
packaging==21.2
paddle2onnx==0.8.2
paddlehub==2.1.1
paddlenlp==2.0.8
paddlepaddle==2.1.3
pandas==1.0.5
paramiko==2.8.0
Pillow==8.1.1
platformdirs==2.4.0
pre-commit==2.15.0
protobuf==3.19.1
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.11.0
pydantic==1.8.2
pydub==0.25.1
pyflakes==2.3.1
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.1
python-multipart==0.0.5
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
rarfile==4.0
requests==2.26.0
scikit-image==0.14.0
scikit-learn==0.24.2
scipy==1.5.4
seqeval==1.2.2
shellcheck-py==0.7.2.1
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
starlette==0.16.0
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.1
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
virtualenv==20.9.0
visualdl==2.2.1
watchgod==0.7
websockets==9.1
Werkzeug==2.0.2
wincertstore==0.2
zipp==3.6.0
